Move config method to context
