+++ lib/context.js
@@ -1,3 +1,6 @@
+const path = require('path');
+const yaml = require('js-yaml');
+
 /**
  * Helpers for extracting information from the webhook event, which can be
  * passed to GitHub API calls.
@@ -57,6 +60,21 @@ class Context {
    */
   get isBot() {
     return this.payload.sender.type === 'Bot';
+  }
+
+  /**
+   * Reads the plugin configuration from the given file name in the `.github` directory.
+   *
+   * @param {github} github Authenticated GitHub API client
+   * @param {string} owner Owner of the repository
+   * @param {string} repo Name of the repository
+   * @param {string} fileName Name of the configuration file within the `.github` directory
+   * @return {Promise<Object>} Configuration object read from the file
+   */
+  async config(fileName) {
+    const params = this.repo({path: path.join('.github', fileName)});
+    const data = await this.github.repos.getContent(params);
+    return yaml.load(Buffer.from(data.content, 'base64').toString()) || {};
   }
 }
