+++ lib/robot.js
@@ -1,5 +1,3 @@
-const path = require('path');
-const yaml = require('js-yaml');
 const {EventEmitter} = require('promise-events');
 const GitHubApi = require('github');
 const Bottleneck = require('bottleneck');
@@ -119,20 +117,6 @@ class Robot {
     }
 
     return probotEnhancedClient(github);
-  }
-
-  /**
-   * Reads the plugin configuration from the given file name in the `.github` directory.
-   *
-   * @param {github} github Authenticated GitHub API client
-   * @param {string} owner Owner of the repository
-   * @param {string} repo Name of the repository
-   * @param {string} fileName Name of the configuration file within the `.github` directory
-   * @return {Promise<Object>} Configuration object read from the file
-   */
-  async getPluginConfig(github, owner, repo, fileName) {
-    const data = await github.repos.getContent({owner, repo, path: path.join('.github', fileName)});
-    return yaml.load(Buffer.from(data.content, 'base64').toString()) || {};
   }
 }
